Allow optional named parameters in CommandLine.pm
Specified with foo=s? get the passed value if any, empty string
otherwise. So --foo will set %options<foo> to '', --foo=bar will result
in %options<foo> := 'bar'. In the first case, %options<foo> will still
be false, but the existance of the argument can be checked with
pir::exists()