<[...-]> without panic
"Unsupported use of - as character range;in Perl 6 please use .. for
range" panic fires even when '-' is last or first character in
the range. This change avoids panic, but only on one case <[...-]>.
Case where '-' is the first thing waits for implementation of <!after >.

Eplain what to do in the panic message.