TODO regex test regressions from -rx
We know these tests are going to fail, so let's just mark that to be
the case.  They're marked regression so that we know which ones to
focus on.

One test description got altered since using # for the issue number
prevented prove from seeing the # TODO later in the line.