fix: kfreebsd build
GNU/kFreeBSD uses the GNU libc whereas Android uses
its own libc. So OS_Linux is not reliable enough to
identify a glibc.

We use __GLIBC__ to know if we are on a glibc instead of
OS_Linux.

__GLIBC__ is defined by features.hi which is pulled
by the glibc headers. That's the reason why we can't
check __GLIBC__ at the begining of the file.