Handle genericity of NFAs for regexes in roles.
Make it so that if you have a regex that mentions a variable that is
a role parameter, then a generic version of the NFA will be saved. As
the regex is incoprorated into the concrete role, this generic NFA is
turned into a concrete one with the value of the role parameter being
substituted in.	Thus it participates fully in LTM.