[src/QAST/*.nqp] shallow-clone nodes with kids
A node without children can make do with the
pir::repr_clone__PP call that was already there.
But those that use the array of children need that
cloned too, so we do self.shallow_clone();

This ought to fix a number of weird quasi-related
behaviors we've seen in Rakudo.