A further round of MOP code simplifications.
We've relied in various places on methods that Parrot's PMCs provide,
but that it's not reasonable to expect a ground-up 6model runtime to
offer. In general, the only things that define internal methods are
the KnowHOW and KnowHOWAttribute, from which everything else is built.
By keeping the MOP simpler, it means we can keep the runtime leaner
and write more of NQP in NQP. This isn't the entire set of changes we
need, but many of them.