Provide nqp:: ops for low-level iterator usage.
NQP will continue to support the $_.key/$_.value approach for hash
iteration in normal code. However, off Parrot this will be provided
through the NQPCORE.setting, and thus is not available to the MOP.
Essentially, these ops are to saw a circularity that we got away with
up until now, but isn't really sensible to burden other backends with.