Add compreg and [get|set]_hll_global abstractions.
We're not going to provide full on namespace support in the nqp:: ops.
However, we will provide languages with a global place to stash stuff.
They can hang their idea of namespace off that or any other useful
things (for example, NQP and Rakudo use it to stash their module
loader).