Refactor how get_attribute_ref works.
Now, get_attribute_ref works more like at_pos_ref, by way of
get_attribute_stable (a new function), so that decoding pointers to embedded
type is delegated to the relevant type, rather than done in nqp.ops. This
eliminates the need for get_attribute_ref to return the number of bits in the
type to the caller via an inout parameter.

Also, this commit temporarily breaks sized int support in CStruct.