Better getting/setting of natively typed attributes.
This brings natively typed attributes more in line with the approach used in
NQP-JVM. Instead of get_attribute and bind_attribute dealing directly with
pointers to the data, we use a NativeValue struct to pass values. This way,
the class can use the attribute's REPR to get and set values.