Start of new HLL-interop infrastructure.
Builds upon the Parrot HLL identifiers. This mechanism will be used to
fix various Rakudo issues when NQP-defiend objects are encountered. It
will also replace the special-case perl6ize_type with a more general
and more portable mechanism, that can work in an environment where all
objects are 6model objects.