Prepare for some extra :decl(...) possibilities.
These will cover things that today require code generation or are not
available at NQP level, but probably should be. It will allow Perl 6's
custom lexpad type to be eliminated in favor of the shared NQP one, as
well as simplifying code generation for certain constructs and giving
us a freer hand in how we compile time. Mostly, this just documents
the new things.