Add a local_lifetime hint op
This tells the code generator that the named locals don't need to escape the
scope of the op.  Note that this is more than just a scope analysis, since it
affects behavior in loops: you're telling the codegen that it's OK to clobber
the named variable between iterations.  As such, to automatically generate it
would require a dataflow analysis.  A future dataflow analyzer might do so.