Fix handling of generic method names.
On JVM (and MoarVM, for that matter) the name associated with a code
object is stored statically, not in the actual code ref itself like on
Parrot. This is generally more efficient (saves a field in closures),
but we relied on it. This prevented declaring more than one user
defined operator in Rakudo on JVM, which busted various tests. On
Parrot, we keep the same approach as today.