Update the STable of delegates in P6Opaque.change_type.
This fixes a bug where cloning a P6Opaque resulted in an object of the wrong
type if the opaque had a delegate and was then change_type-d without needing
to introduce a new delegate. Thus obj.st != obj.delegate.st. Since
P6OpaqueBaseInstance.clone simply clones the delegate if one is present, this
resulted in an object of the wrong type under these conditions.