[NQP::Grammar] Remove redundant regex line
Since rule unv contains \h+ as a possible alternative, <.unv>? in the
second (and surviving) line of ENDSTMT is sometimes [\h+]?, which is
functionally equivalent to \h*.

The first (and now removed) line of ENDSTMT was essentially line 2 with
<.unv>? replaced by \h*. As stated <.unv>? already covers what \h* does,
so line 1 was removed. No regressions or speed drops result from the
removal of this line.