Optimize when regex calls back to the main lang.
This invokes the optimizer of the main language when there's a call
back into it. This in turn means we don't need to poison all lowering
of lexicals in the NQP optimizer in token/rule/regex. For now, we will
poision if we spot a call back into the main language; this can be
relaxed later.