implement nativecallcast op for parrot backend
The op itself seems to work for all kinds or target types, but there
is an issue with nums and strs in NativeCall when we try to box the
return value of the op. Native ints work out for some reason even
when the code paths seem to be identical.