RT #67122 fix slowdown in large character ranges in regexes
On all platforms it built a string with every codepoint of the given range,
and when :i was in effect it built that string with uppercase and lowercase
codepoints to do a "cheap" character lookup later. Building this string is
very expensive and might even crash on long ranges.
Now it just checks if a char is within the range, (lower <= char <= upper).

Benchmark: http://froggs.de/perl6/charrange.html