allow code block in ** quantifiers, / foo ** { $a := 42 } /
In NQP the block can either provide a single item, which would be
handled as foo**42, means that it must match exactly $item times.
If the block provides a list [$min, $max], then the first two elems
are used, like in foo**0..3.
Note that the behaviour in rakudo is different, since we have Range
objects there.