Map nqp::objectid(...) op.
On Parrot, it maps to what nqp::where(...) does and can continue to,
since Parrot doesn't move objects. On MoarVM it maps to the new VM op
of the same name. On JVM, it currently maps to nqp::where, but this
will want reviewing as JVMs can also move objects.