Avoid a bunch of str boxing in !cursor_pass.
When we have a conditional in non-void context, it needs to produce a
single kind of result; since a method call cannot be known to unbox to
a native string, then this caused us to perform a boxing operation.
This gets rid of a boxing operation on every !cursor_pass.