Various micro-optimizations to MATCH.
These avoid various bits of boxing. In fact, all of the boxing that
we accidentally did in MATCH. In a parse-heavy benchmark such as the
NQP JSON parser, this cuts the number of object allocations roughly
in half by number (though since it cuts down on boxing, and boxes are
comparatively small, it's a much smaller decrease on GC runs needed,
from 29 to 25).