Revert Parrot to 6.10.0
During the release process of 2015.02, (which will be the last Rakudo compiler
release on parrot for quite some time to come), it became clear that the
Parrot 7.x series does *not* build on Windows.  In order to not hold up the
Rakudo release process, it was decided to revert the Parrot version to 6.10.0,
the last version known to build on Windows *and* pass the Rakudo sanity test
as well as the spectest.