optimize after of known length to before w/ offset
This saves allocating and flipping a million or so bytes in rakudo compilation
just for the occasional <?after '>'> and such.  In fact, all of the afters in
nqp and rakudo are of known length, though one of them is just <?after .>.
We handle both dot and character classes, as well as any embedded anchors.