The integer serialization tests now "pass" on 32 bit Parrot, albeit with TODOs.
commit 1a559ff0fd3ba34c in Feb 2014 added "Serialization tests for integers
up to 2**62." As written, they fail on 32 bit Parrot, because unlike all other
platforms, there NQP's native integers are not 64 bit. Specifically, the way
the looping was written too few tests would be run, meaning that the failure
was only detected because the test count didn't match the plan.

Refactor the looping construction so that the correct number of tests are
always run (on all platforms). The tests aren't *meaningful* on 32 bit Parrot
because they aren't testing that integers >= 2**31 can serialize, but they
don't fail (because as before they instead end up testing integers around 0,
rather than 2 ** $n, for $n in 31..62)

Add a test, TODO'd on 32 bit parrot, that the values being tested are actually
large and positive (which 0 obviously isn't).

With this, we have no broken windows, and all tests (usefully) pass on MoarVM,
JVM and 64 bit Parrot.