allow use of cstack for pos-only advancement
fixes RT #125345
The cstack is used not only for named captures, but occasionally the top
element is needed for its position, even if the capture is anonymous.  For
now, anyway, when that situation arises, just make sure there is at least
one anonymous capture on the stack to carry the position.  (But two or more
anonymous captures can be coalesced, reducing load on method MATCH.)