Remove wrong and duplicate optimization.
We don't know that the atom we're looking at really can cope with us
swallowing multiple chars into the literal at this point; we may be
in a goal or a separator. Rather than fix it, it can simply go away,
since there's code in the termish action method that does the very
same optimization, but more generally, and when we know it's safe. So
we don't lose out, but we do lose a couple of weird bugs.