throw out nqp::while return value
it used to return the value of the condition, but that
doesn't help rakudo with its semantics. instead, it caused
every condition of a while loop to have .sink attempted
on it, even in cases where we know it doesn't make sense.

so now we either have the register be an object that's null'd
or set the return kind to void.

Also, remove fossilized IMM_ARG mention.