cleanup handling of unkown ops
We used to catch an exception to append source information to the message
in order to die again using that more informative message. Now we just
put this extra information in the very first exception. That helps us to
avoid a problem where we visited the same CATCH block twice.