Fix getBI and makeBI again, differently this time.
For Rakudos IntStr, we have to rely on the target type to get the right
unboxIntSlot, because for some reason the LHS already comes with the same REPR
but different REPRData. It works for now, but I can imagine cases where it
would stop working - I'm not sure how likely these cases are to occur, though.