Remove useless conditional in sprintf
Unless I've severely forgotten my math, a < b && a > b  is never True.
This ternary inside an if statement was effectively testing that,
meaning the False portion would always happen. Doesn't make a difference
to remove this ternary, but I think it's nice to not be using a
conditional that isn't very conditional :) .