Retain newlines in continued code
Say we want to define a class in our REPL, we might input:

    > class Foo {\
    * method foo {...}\
    * method bar {...}\
    }

Only to see

    ===SORRY!=== Error while compiling <unknown file>
    Strange text after block (missing semicolon or comma?)
    at <unknown file>:1
    ------> class Foo {method foo {...}⏏ method bar {...} }
        expecting any of:
            infix
            infix stopper
            postfix
            statement end
            statement modifier
            statement modifier loop

Which is surprising, and surprises are bad.

So, after we chomp the input in `deadline`, we unchomp it if we get a continuation, which means that the compiler sees code that looks like:

    class Foo {
    method foo {...}
    method bar {...}
    }

which is legal, rather than code like `class Foo { method foo {...}method bar {...}}`, which isn't. Everybody wins.