Avoid a load of hash allocation/copies.
Normally we just set symbol properties once ever, so just use the
slurpy hash we already have rather than creating a new one and copying
to it. Saves a ton of hash and hash iterator allocations.

This makes `for ^500 { EVAL 'regex { abcdef }' }` in Perl 6 run in
about 85% of the time it used to (and will similarly help other EVAL
heavy things, and shave a bit off compile times generally).