Support for nested compilation
EVALed code is considered an independent compilation unit that gets compiled
using a new World which contains an independent serialization context. It also
gets its own QAST compiler.

When a BEGIN time EVAL is run during precompilation of a module, the
independent SC would not be written into any file but would still be referenced
from the outer compilation unit. We therefor use the outer world's SC. We also
need to share a couple of other attributes that keep track of the compilation's
state.

For easier sharing, these attributes are kept in a CompilationContext object.
The QAST compiler's frames can be shared by making it possible to pass in the
hash where those frames are stored.