Make HLL::Grammar.O take named arguments.
This fixes the problem of the method manually parsing a string (#145). Since
the old parser code allowed a construct like '%additive, :op<add>' to copy
values from %additive and set the 'op' key to be 'add', I've added a key
'inherit' which starts out with the named information and then writes in the
named arguments.