[JVM] Avoid constant pool entries for large ints
Rakudo does not build on the JVM backend currently, because
the limit for constant pool entries (65535) is reached.

We have a lot of integers > 32677 (max value for type Short)
in the constant pool. Most (if not all) of them are added
when nodes of type QAST::WVal are processed in method 'as_jast'
(in src/vm/jvm/QAST/Compiler.nqp) and nqp::scgetobjidx returns
those large indices.

This commit avoids using the constant pool for pushing
integers > 32677 to the stack. Instead we increase the value
on the stack via 'iadd' until the required value is reached.

If we find a cleaner way to work around the limit for constant
pool entries, this commit should be reverted.