Uniqueify fates in prep for NFA optimizations
Prior to now, fates were always numbered starting with 1 for protos
and starting with 0 for alts.  We are moving to a model in which all
fate numbers are unique across the process.  Every NFA now knows its
own range of fates, in the range of fbeg ..^ fend.  Conveniently, for
both protos and alts, the 0th fate name entry was already unused and
set to 0, so we encode the new fbeg and fend into the 0th fate name.
Eventually, the uniqueness of fates will allow us to use a top-level
NFA scan to determine in advance exactly what some of the sub-scans
will do.  Currently we have duplicate scans for lower-level rules;
when the top-level scan can return multiple levels of fates to the
sub-NFAs, we'll be able to skip some of those redundant NFA scans.