unbreak builds on jvm, js
Turns out the first approach won't work on jvm because of fatename[0] isn't
initialized to 0 in the bootstrap.  So we'll use NFA attributes, and process
fate offsets in a manner that should work without the need to patch each vm.
(For now, though, the new mechanism is enabled only for proto, not alt.)