Fix sprintf precision parameter for `d`
Currently, it just chops the number to $precision digits, producing
wrong results. The correct behaviour is for the precision to signify
how many digits long the number should be, prefixing it with zeros if
needed. This also results in an empty string if precision is zero and
the number is zero too.

Bug find: https://irclog.perlgeek.de/perl6-dev/2017-01-22#i_13966753