start to unify NQPMatch with NQPCursor
This will eventually make it possible to cut down all the copying
currently done by the MATCH method.  We have not quite got rid of
the old match object yet, but it is now reduced to a stub that knows
how to return true under Bool.  (Haven't figured out how to make it
go away entirely in favor of a simple boolean flag...)