finish s/NQPCursor/NQPMatch/ tranformation
NQPMatch is now both a cursor and a match object.  (The name 'NQPCursor'
still exists but is an alias to NQPMatch now.)  A new method is provided
to clone match information for when we need a new NQPMatch identity for
operator reductions.  (The rakudo compiler also uses this for its old
'AS_MATCH' method.)

The )> regex assertion is now passed on the cstack as '$!pos' rather than
as '$!to', since at that level NQPMatch is more like a cursor than a match.

Now passes all nqp tests again.