separate the $!to concept back out from $!pos
This will allow )> to work correctly again. It may also give us the
opportunity to optimize away <.ws> cursor creation by just tacking it onto
the previous capture cursor as a $!pos update that doesn't affect $!to.
This in turn will cut down on $!cstack usage and scanning overhead. In
some cases it may allow us to steal the cstack directly for a match list,
without having to grep out the non-capture cursors.

Also did some Num suppression on subtractions that might not know their
operand types and default to Num.