Fix for RT #130760 - print usage instructions to stderr when an invalid command line option is used
Entering an invalid option when invoking nqp or perl6 leads to an
error message being printed followed by usage instructions. In both
cases this information is written to stdout when it should have been
written to stderr. This fix solves the problem for nqp by introducing
a flag which tells the code doing the writing which fd to use.
In order to fix perl6 a separate Rakudo PR is necessary. That fix
will be dependent on this one but it won't break anything even if this
fix is reverted, it will just continue to work the way it does today.