Allow for extra config info to be shown from HLL
This uses the approach of a dynamic variable that, when called, will
nqp::say whatever it thinks it needs to say as extra config information.

I tried just calling a subroutine: that also works, but spoils the global
namespace even further: a dynamic variable only needs registration in a
hidden hash.  I also tried returning the value, but that results in a
big fail because it somehow warns and there is no warn handler installed,
so it bombs.

Should one change this, then this would also need changing in
src/core/Process.pm .