[unicode-grant] Update QASTRegexCompilerMAST to use new indexicim and eqaticim ops
This awesomely cleans up the code for handling the different combinations
of ignorecase and ignoremark. Previously they were each handled slightly
differently, which can easily create bugs between the different implementations.

This also fixes RT #128875 which was a bug where ignorecase+ignoremark
would give false positive matches if the first grapheme of the needle matched
any of the graphemes in the haystack.