Add eqatim and indexim ops. Fix a bug when using ignoremark
There was a bug with ignoremark if the first two codepoints in the haystack
matched, it would act as if the entire needle had matched. This was only with
ignoremark by itself, as ignorecase+ignoremark had already been fixed.

This bug was less noticable than the previously fixed bug in
ignorecase+ignoremark, but it has been solved the same way as the
ignorecase+igoremark had been solved — adding new MoarVM ops.

In this case there already was a nqp::eqatim op, but it did not work properly.
This op has been fixed upstream in MoarVM in addition to the nqp::indexim op
being added.