Add new --ignore-errors flag to ./Configure.pl
Will allow you to compile nqp even if MoarVM/JVM versions
are not high enough. May ignore certain other errors as well.
It could be extended to allow functioning when other errors
occur, but that will need to be added as needed, since the
script often doesn't pass information along when failing
conditions are encountered.