[jvm] Die with 'Division by zero' in div_i
Without this change a division by zero led to a Java exception:
java.lang.ArithmeticException: / by zero

For some reasons that didn't work well with 'dies-ok'. As a
result t/nqp/059-nqpop.t was blowing up on the jvm backend.
With this patch 'make test' is clean, again. Also, the error
message is now identical to the other two backends.