Don't die if parent doesn't have dump method
The dump_extra_node_info method of the QAST::SpecialArg role would look
up the method of the same name on the immediate parent and invoke it,
dying if that method didn't exist and could not be invoked (on
QAST::Want, for instance). Modifying the behavior so that it checks
whether the there is an invokable method, and if not continue along the
parent chain until one is found or all are exhausted.