Make certain conditionals with natives 2.1x faster
Addresses some cases in https://github.com/rakudo/rakudo/issues/1416

getlexref is expensive compared to plain getlex, so if we know
we're not gonna use the conditional, try coercing it to a simpler value.
This gets rid of lexref and if we got native-type value, goes to use
the native if_/unless_ op.

The perf improvement applies only nqp::if/nqp::unless/nqp::with/
nqp::without when the conditional is a variable and the conditional is
not used for anything other than deciding the branch. I would assume other
cases would've be optimizeable, but I kept getting confusing crashes
during nqp compilation when I tried.