[jvm] Allow 'chain' op to use child as callee
Modify the 'chain' op to allow the option to use the first child as the
callee. Before the name of the op served as the operator sub and the
children the operands. This modification makes it so that, if there is
no name provided to the chain, child 0 serves as the operator and
children 1 and 2 the operands.

This modification is being made to coincide with a Rakudo development
allowing negated chained ops to continue to work as chained.

See <https://github.com/rakudo/rakudo/issues/1304>.