teach old html frontend a bit about multiple threads
still missing: call graph for the rest of the threads,
GC tab ought to display a bit of information about all
the other participants (if applicable), and there's
a bug causing incredible amounts of promoted bytes in
the GC list.

Also, a decision on whether "inclusive" time is allowed
to go past 100% or if everything should be scaled down
to "total cpu time".