Fix concurrency failure in threadyield sibling test
Pushing onto an array is not an atomic operation. The array's storage may have
to be expanded first before the new element can be added. Thus a simple check
for nqp::elems(@a) is not enough to ensure that the push is already done. We
also have to check for the actual end result before we may continue. Otherwise
two push operations may run concurrently and one of the updates may be lost.