[JVM] Adjust 'getrusage' to fix rakudo build
The api for getrusage has been changed -- it takes an argument now.[1]

Also, this commit adds a workaround for non-working native arrays
in Rakudo's ThreadPoolScheduler.[2] Declaring @rusage as a native int
array comes out as a VMArrayInstance_i instead of a VMArrayInstance.
This really needs a clean fix, but I was unable to do that.

[1] https://github.com/rakudo/rakudo/commit/c71fa17d8d
[2] https://github.com/rakudo/rakudo/issues/1666