Invert control flow of build process to support parallelization
Before this change we had an outer most loop pushing new build targets
through the build code and a tiny inner most loop basically waiting for
a command to finish. To support parallelization, we have to invert this.
Now the outer most loop is the event loop waiting for commands to
finish. Whenever needed it will pull a new build job from the dependency
tree.
Unintuitively the most efficient way to find the next job is to traverse
the dependency tree anew every time. As we only traverse branches that
are not yet marked done, we will find a job to run in very few steps.