Implement `cvalue` method for [NIS]Val QAST Nodes
A lot of the values we set to these nodeas are compile-time known,
and this is a convenience method to mixin QAST::CompileTimeValue
and set $!value/$!compile_time_value to the given value, right from
the constructor.

Originally, I tried `does` QAST::CompileTimeValue to these nodes,
like we do with QAST::WVal, but NQP compilation would blew up.