JVM: Use <surrogate> <private-use> <unassigned> and <reserved> for cp names
This fixes the return for > 0x10FFFF codepoints so it returns
<unassigned> like it does on MoarVM. It also returns <surrogate-XXXX>
for surrogates instead of the JVM's normal output which is not the same
as we now do for MoarVM fitting with the Unicode recommendations.

This also returns <private-use-XXXX> for private use areas.
Returns <reserved-XXXX> for codepoints that have don't have a name yet
in the Unicode version used by the running JVM.