Fix rebootstrap broken by SC name collision
Since serialization context IDs are derived from the source code, the compiler
ended up overwriting its own SC entries whith the compiler it's compiling.
Break the circularity by extending the SC ID for compilation of stage 1.
This way stage 0 is different from stage 1 which is different from stage 2
which ends up as stage 0 on a rebootstrap.