Modified JVM getsignals to use sh kill builtin
Modified the JVM backend's retrieval of signal values. The problem was
that the '/bin/kill -l'  was inconsistent among different
implementations and producing incorrect results in some. For example, on
Fedora 27, '/bin/kill -l' produces:

HUP INT QUIT ILL TRAP ABRT IOT BUS FPE KILL USR1 SEGV USR2 PIPE ALRM
TERM STKFLT CHLD CLD CONT STOP TSTP TTIN TTOU URG XCPU XFSZ VTALRM PROF
WINCH IO POLL PWR UNUSED SYS

Clearly, this method of determining the signum by sequence in the
'/bin/kill -l' output is flawed. This could be accounted for if this
output was consistent across platforms, but it is not.

Using the Bourne shell (sh) kill built-in as that seems to be the most
consistent and portable across all of the POSIX-complient systems tried.
Using the following form to retrieve the name given a signal value:

    sh -c "kill -l $signum"

While this spawns more processes, it allows us to be sure that the
signals names and values match and does not have the same potential for
error that the previous method had.