Disable the SC write barrier during load
The load time of a module is dynamically within the scope of the
compilation of the using module. However, from a user perspective the
module mainline is probably more fittingly seen as a bit of runtime
during compile time.

So, we'll try it like this. The immediate motivation is the MinG issue
in https://github.com/rakudo/rakudo/issues/2056. Its `our`-scoped
variables that are exported never used to have their assignments made
subject to precompilation repossession since Scalar's assignment didn't
use the SC write barrier. Now, Scalar assignment uses `bindattr` and so
is subject to the SC write barriering. That would count as a potential
bug fix in the case of meta-objects implemented in Perl 6 being used
with `augment` (not that any known cases of this exist). However, it
has the unfortunate consequence that assignments made in a module's
mainline now trip the barrier and end up serialized into the using
module's serialization context when a variable was exported.

This change together with the recent Rakudo change to make Scalar less
of a special case have the net effect of preserving existing behavior
of exported variables. It probably also makes the actual behavior of
code in a module's mainline a bit more runtime-y, and thus may well
help to make precompilation a little more transparent than it is today
also. Hopefully, with no downsides.